# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OnlinePort",
        "version": "0.1"
    },
    "paths": {
...
        "/api/order-statuses": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of Order Statuses",
                "operationId": "3a002ba5924af70e7b8f3e1440fba633",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Store Order Status",
                "operationId": "74d93fe985b453247fc8e328f1133729",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "OrderStatusNameTest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
        "/api/order-statuses/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show Order Status",
                "description": "Show Order Status",
                "operationId": "2f5c50af4acd67e93ceae460d94c6fec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update Order Status",
                "description": "Update Order Status",
                "operationId": "Update Order Status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "OrderStatusNameTest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Summary",
                "description": "Orders",
                "operationId": "Delete Order Status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
...
    }
}